backend/requirements.txt
numpy==1.13.3
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
urllib3==1.22
web.py==0.38
h5py
tensorflow
keras
jieba
gensim